requirements.txt
# === Core Framework ===
django>=5.2,<6.0
djangorestframework
djangorestframework-simplejwt

# === Deployment ===
gunicorn==20.1.0  # WSGI HTTP server for production

# === Database ===
psycopg2-binary  # PostgreSQL adapter

# === Environment Management ===
python-dotenv  # Load environment variables from .env files

# === API & Authentication ===
PyJWT  # JWT token handling
requests  # For making HTTP requests (e.g., to third-party services)

# === Middleware ===
django-cors-headers  # Handle CORS (Cross-Origin Resource Sharing)
